Fix per comments
