+++ solr/core/src/java/org/apache/solr/core/SolrCore.java
@@ -70,8 +70,6 @@ import org.apache.lucene.index.IndexDeletionPolicy;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.search.CancellableCollector;
-import org.apache.lucene.search.CancellableTask;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.store.IndexInput;
@@ -178,7 +176,6 @@ import org.apache.zookeeper.data.Stat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.apache.solr.common.params.CommonParams.CUSTOM_QUERY_UUID;
 import static org.apache.solr.common.params.CommonParams.PATH;
 
 /**
@@ -202,9 +199,7 @@ public final class SolrCore implements SolrInfoBean, Closeable {
    */
   public final UUID uniqueId = UUID.randomUUID();
 
-  //TODO: This needs to become a time aware storage model
-  private final Map<String, CancellableTask> activeCancellableQueries = new ConcurrentHashMap<>();
-  private final Map<String, String> activeQueriesGenerated = new ConcurrentHashMap<>();
+  private final CancellableQueryTracker cancellableQueryTracker = new CancellableQueryTracker();
 
   private boolean isReloaded = false;
 
@@ -3252,73 +3247,8 @@ public final class SolrCore implements SolrInfoBean, Closeable {
     return blobRef;
   }
 
-  /** Generates a UUID for the given query or if the user provided a UUID
-   * for this query, uses that.
-   */
-  public String generateQueryID(SolrQueryRequest req) {
-    String queryID;
-    String customQueryUUID = req.getParams().get(CUSTOM_QUERY_UUID, null);
-
-    if (customQueryUUID != null) {
-      queryID = customQueryUUID;
-    } else {
-      queryID = UUID.randomUUID().toString();
-    }
-
-    if (activeQueriesGenerated.containsKey(queryID)) {
-      if (customQueryUUID != null) {
-        throw new IllegalArgumentException("Duplicate query UUID given");
-      } else {
-        while (activeQueriesGenerated.get(queryID) != null) {
-          queryID = UUID.randomUUID().toString();
-        }
-      }
-    }
-
-    activeQueriesGenerated.put(queryID, req.getHttpSolrCall().getReq().getQueryString());
-
-    return queryID;
-  }
-
-  public void releaseQueryID(String inputQueryID) {
-    if (inputQueryID == null) {
-      return;
-    }
-
-    activeQueriesGenerated.remove(inputQueryID);
-  }
-
-  public boolean isQueryIdActive(String queryID) {
-    return activeQueriesGenerated.containsKey(queryID);
-  }
-
-  public void addShardLevelActiveQuery(String queryID, CancellableCollector collector) {
-    if (queryID == null) {
-      return;
-    }
-
-    activeCancellableQueries.put(queryID, collector);
-  }
-
-  public CancellableTask getCancellableTask(String queryID) {
-    if (queryID == null) {
-      throw new IllegalArgumentException("Input queryID is null");
-    }
-
-    return activeCancellableQueries.get(queryID);
-  }
-
-  public void removeCancellableQuery(String queryID) {
-    if (queryID == null) {
-      // Some components, such as CaffeineCache, use the searcher to fire internal queries which are not tracked
-      return;
-    }
-
-    activeCancellableQueries.remove(queryID);
-  }
-
-  public Iterator<Map.Entry<String, String>> getActiveQueriesGenerated() {
-    return activeQueriesGenerated.entrySet().iterator();
+  public CancellableQueryTracker getCancellableQueryTracker() {
+    return cancellableQueryTracker;
   }
 
   /**
